+++ great_expectations/profiler/expectation_configuration_builder/default_expectation_configuration_builder.py
@@ -7,8 +7,8 @@ from great_expectations.profiler.expectation_configuration_builder.expectation_c
 )
 from great_expectations.profiler.parameter_builder.parameter_container import (
     ParameterContainer,
+    get_parameter_value,
 )
-from great_expectations.profiler.util import get_parameter_value
 
 
 class DefaultExpectationConfigurationBuilder(ExpectationConfigurationBuilder):
