+++ great_expectations/profiler/parameter_builder/multi_batch_bootstrapped_metric_distribution_parameter_builder.py
@@ -12,8 +12,8 @@ from great_expectations.profiler.parameter_builder.parameter_container import (
     DOMAIN_KWARGS_PARAMETER_NAME,
     ParameterContainer,
     build_parameter_container,
+    get_parameter_value,
 )
-from great_expectations.profiler.util import get_parameter_value
 from great_expectations.validator.validation_graph import MetricConfiguration
 from great_expectations.validator.validator import Validator
